fix: move file to servises/api
